Implements chat page and initial matching algorithm
Adds a chat page that validates the existence of a chat session and fetches chat messages.

Integrates realtime chat component and implements logic to validate
if the session exists in the database and is active. If the session
is invalid, the user is redirected to the home page. Additionally,
fetches existing chat messages for the session and displays them.

Shortens chat session expiration time for testing purposes.
Removes the user from the waiting queue after session creation.